server/requirements.txt
eventlet
Flask
Flask-SocketIO
python-dotenv